new safe_pypsa_import for pnl attrs that were emptied on export
If all values for a pnl[attr] are 0., pypsa exports an empty frame,

see

https://github.com/PyPSA/PyPSA/blob/ebda867913f04c132a4fe20b9fccbeb8f503103c/pypsa/io.py#L526

This confuses the code when reading in the results - they are not
imported as 0s but as an empty frame.

So this new function safe_pypsa_import fills in the zeros.

Also in plot_network a new check that demand and supply balance in
every hour.