add term to objective function to value battery SOC at end of day
This prevents the simulation from draining the battery unnecessarily,
especially when the simulation is not extended after 24 hours.

The default is to value it at 90% of the cost of the H2 turbine, so
that H2 turbine is displaced as priority, then the battery is filled.

Also make changes to allow extended_hours = 0, since there is no need
to extend any more.

Finally: correct name of Future-Agorameter.